Introduce button functionality
The program has been updated to take button inputs to
initiate a round, instead of playing five rounds by
typing in an alert window.

Results are logged in the console.

The quotation marks have been removed around the
interpolation of lose- and winStrings for consistency.

The game() function is unused for now; commented out.